requirements.txt
pyyaml~=5.4.1
click~=7.1.2
requests~=2.25.1
tabulate~=0.8.7
jinja2~=2.11.3

pytest~=6.2.2
pep8~=1.7.1
